refactoring according PR
